Refactoring Publish schools search controller
