+++ app/views/publish/providers/schools/search/new.html.erb
@@ -22,7 +22,7 @@
     <div class="govuk-form-group<% if f.object.errors.present? %> govuk-form-group--error<% end %>">
       <%= f.label :query, { class: "govuk-label govuk-label--l", for: "publish-schools-search-form-query-field" } do %>
         <span class="govuk-caption-l">Add school</span>
-        <%= I18n.t("publish.providers.school_search.new.title") %>
+        <%= t(".title") %>
         <% if f.object.errors.present? %>
           <span class="govuk-error-message" id="publish-schools-search-form-query-field-error" data-qa="provider-error">
             <%= f.object.errors.first.message %>
