+++ app/views/publish/providers/schools/search/results.html.erb
