+++ config/locales/en.yml
@@ -647,6 +647,11 @@ en:
         address4: "County"
         postcode: "Postcode"
       schools:
+        search:
+          new:
+            title: &new_search Enter a school, university, college, URN or postcode
+            caption: "Add school"
+            cannot_find: &cannot_find I cannot find the school - enter manually
         index:
           add_school: &add_school "Add school"
         new: *add_school
@@ -662,11 +667,6 @@ en:
         town: &school_town "Town or city"
         address4: &school_county "County"
         postcode: &school_postcode "Postcode"
-      school_search:
-        new:
-          title: &new_search Enter a school, university, college, URN or postcode
-          caption: *add_school
-          cannot_find: &cannot_find I cannot find the school - enter manually
       study_site_search:
         new:
           title: *new_search
